+++ src/main/scala/Main.scala
@@ -57,45 +57,16 @@ object Main extends App {
   }
 
   val inp =
-    """ DEPRECATED
-      | REMOVED - replaced by SwiftUI
-      | ADDED
-      | MODIFIED
-      | UPCOMING
-      | ALAMAKOTA a kot ma Ale
-      | Construct and manage a graphical, event-driven user interface for your
-      | iOS or tvOS app.
-      |
-      | The UIKit framework provides the required infrastructure for your iOS or
-      | tvOS apps. It provides the window and view architecture for implementing
-      | your interface, the event handling infrastructure for delivering Multi-
-      | Touch and other types of input to your app, and the main run loop needed
-      | to manage interactions among the user, the system, and your app. Other
-      | features offered by the framework include animation support, document
-      | support, drawing and printing support, information about the current
-      | device, text management and display, search support, accessibility
-      | support, app extension support, and resource management.
-      |
-      | ! Important
-      |   Use UIKit classes only from your app’s main thread or main dispatch
-      |   queue, unless otherwise indicated. This restriction particularly
-      |   applies to classes derived from UIResponder or that involve
-      |   manipulating your app’s user interface in any way.
-      |       def Maybe a
-      |           sub x y = x - y
-      |           def Nothing
+    """Funkcja ()
+      |  x <- 1
+      |  return x
       |""".stripMargin
 
-  /** Invoking the Enso Documentation Parser */
-  println("===== DOCUMENTATION =====")
-  val documentation    = new DocParser().runMatched(inp)
-  val htmlPath         = "target/"
-  val cssFileName      = "style.css"
-  val title            = "Foo Bar Baz"
-  val fileName         = "output"
-  val documentationHtml = DocParserHTMLGenerator.generateHTML(documentation, htmlPath, cssFileName, title, fileName)
-  println(pretty(documentation.toString))
+  /** Invoking the  Parser */
+  println("===== PSEUDO PARSER =====")
+  val parsed    = new Parser().runMatched(inp)
+  println(pretty(parser.toString))
   println("------")
-  println(documentation.show())
+  println(parser.show())
   println("=========================")
 }
